move currentValuesChanged subscription into PlotLegend
