+++ packages/studio-base/src/panels/Plot/Plot.tsx
@@ -595,28 +595,13 @@ export function Plot(props: Props): JSX.Element {
     }
   }, [coordinator, setGlobalBounds, shouldSync]);
 
-  const [currentValuesBySeriesIndex, setCurrentValuesBySeriesIndex] = useState<unknown[]>([]);
-  useEffect(() => {
-    if (!coordinator || !config.showPlotValuesInLegend) {
-      return;
-    }
-    const handler = (values: unknown[]) => {
-      setCurrentValuesBySeriesIndex(values);
-    };
-    coordinator.on("currentValuesChanged", handler);
-    return () => {
-      coordinator.off("currentValuesChanged", handler);
-      setCurrentValuesBySeriesIndex([]);
-    };
-  }, [coordinator, config.showPlotValuesInLegend]);
-
-  const valuesBySeriesIndex = useMemo(() => {
+  const hoveredValuesBySeriesIndex = useMemo(() => {
     if (!config.showPlotValuesInLegend) {
       return;
     }
 
     if (!activeTooltip?.data) {
-      return currentValuesBySeriesIndex;
+      return;
     }
 
     const values = new Array(config.paths.length).fill(undefined);
@@ -625,12 +610,7 @@ export function Plot(props: Props): JSX.Element {
     }
 
     return values;
-  }, [
-    activeTooltip,
-    config.paths.length,
-    config.showPlotValuesInLegend,
-    currentValuesBySeriesIndex,
-  ]);
+  }, [activeTooltip, config.paths.length, config.showPlotValuesInLegend]);
 
   const { keyDownHandlers, keyUphandlers } = useMemo(() => {
     return {
@@ -676,6 +656,7 @@ export function Plot(props: Props): JSX.Element {
         {/* Pass stable values here for properties when not showing values so that the legend memoization remains stable. */}
         {legendDisplay !== "none" && (
           <PlotLegend
+            coordinator={coordinator}
             legendDisplay={legendDisplay}
             onClickPath={onClickPath}
             paths={series}
@@ -683,8 +664,8 @@ export function Plot(props: Props): JSX.Element {
             saveConfig={saveConfig}
             showLegend={showLegend}
             sidebarDimension={sidebarDimension}
-            valuesBySeriesIndex={valuesBySeriesIndex}
-            valueSource={activeTooltip?.data ? "hover" : "current"}
+            showValues={config.showPlotValuesInLegend}
+            hoveredValuesBySeriesIndex={hoveredValuesBySeriesIndex}
           />
         )}
         {showResetViewButton && (
